+++ src/algorithmTester/AlgorithmTester.java
@@ -5,14 +5,29 @@ import model.parser.Model;
 import servlet.Result;
 import servlet.rest.GET;
 import servlet.rest.REST;
+import util.Logger;
 
 import static marker.AlgorithmImplementation.USER;
 import static servlet.rest.REST.preprocess;
 
 public class AlgorithmTester implements RestEndpoint {
+    ImplementationValidator validator = new ImplementationValidator();
+
     public static void run() {
         REST.start();
-        System.out.println("[INFO] Started AlgorithmTester");
+        Logger.info("Started AlgorithmTester");
+    }
+
+    @GET("/validate/:methodName/:name/:compact")
+    public String validate(String compact, String methodName, String name) {
+        return new Result(() -> validator.validate(methodName, name, Boolean.parseBoolean(compact)))
+                .computeJSON();
+    }
+
+    @GET("/validateAll/:name/:compact")
+    public String validateAll(String compact, String name) {
+        return new Result(() -> validator.validateAll(name, Boolean.parseBoolean(compact)))
+                .computeJSON();
     }
 
     @GET("/isDeterministic/:automaton")