+++ src/main/Frontend.java
@@ -1,13 +1,14 @@
 package main;
 
 import spark.Spark;
+import util.Logger;
 
 import java.io.File;
 
 public class Frontend {
     public static void run() {
         Spark.staticFiles.externalLocation(getStaticFileLocation());
-        System.out.println("[INFO] Started Frontend. Go to: http://localhost:4000");
+        Logger.info("Started Frontend. Go to: http://localhost:4000");
     }
 
     private static String getStaticFileLocation() {