+++ Changelog.md
@@ -19,8 +19,13 @@
 * Select an algorithm to apply on a Finite Automata
 * Code your own implementation and test it
 ---
-### 1.5.5 CTL formula generator
+### 1.5.7 CTL Formula Generator
 * Added CTL formula generator
+### 1.5.6 Created Logger
+* Contains static methods that log to `System.out` even when called in `Result`
+* Added video on how to create and test your own algorithm to frontend
+### 1.5.5 Created Algorithm validation and Report
+* Report is saved in `/resources`
 ### 1.5.4 Created light/dark mode button for graph
 * Clicking the button inverts the graph colors for all graphs
 ### 1.5.3 Added hints to buttons that refer you to tutorials